src/backend/requirements.txt
fastapi
uvicorn
kubernetes
pydantic
pydantic-settings
jinja2
pyyaml
uuid
python-multipart
starlette
sqlalchemy